+++ app/src/main/java/feelings/helper/ui/answer/AnswerActivity.java
@@ -31,7 +31,7 @@ import static java.util.Arrays.asList;
 
 public class AnswerActivity extends AppCompatActivity {
 
-    private int questionId;
+    private long questionId;
     private EditText answerText;
     private ExpandableListView feelingsListView;
     private List<String> feelingsGroups = new ArrayList<>();
@@ -42,7 +42,7 @@ public class AnswerActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.answer_activity);
 
-        questionId = getIntent().getIntExtra(QUESTION_ID_PARAM, -1);
+        questionId = getIntent().getLongExtra(QUESTION_ID_PARAM, -1);
 
         TextView questionView = (TextView) findViewById(R.id.question_text_on_answer);
         questionView.setText(QuestionService.getQuestionText(this, questionId));
